[SPOT-281][REFACTOR] RSA 관련 코드 Auth 도메인으로 이동
