+++ src/main/java/com/example/spot/auth/application/legacy/AuthServiceImpl.java
@@ -13,7 +13,7 @@ import com.example.spot.member.domain.association.StudyJoinReasonRepository;
 import com.example.spot.member.presentation.dto.MemberRequestDTO.SignUpDetailDTO;
 import com.example.spot.member.presentation.dto.MemberResponseDTO.CheckMemberDTO;
 import com.example.spot.member.presentation.dto.MemberResponseDTO.NicknameDuplicateDTO;
-import com.example.spot.member.presentation.dto.rsa.Rsa;
+import com.example.spot.auth.presentation.dto.rsa.Rsa;
 import com.example.spot.auth.domain.RefreshToken;
 import com.example.spot.auth.domain.VerificationCode;
 import com.example.spot.member.domain.enums.Carrier;
@@ -22,7 +22,7 @@ import com.example.spot.member.domain.enums.LoginType;
 import com.example.spot.member.domain.enums.Status;
 import com.example.spot.member.domain.MemberRepository;
 import com.example.spot.auth.domain.RefreshTokenRepository;
-import com.example.spot.member.domain.rsa.RSAKeyRepository;
+import com.example.spot.auth.domain.rsa.RSAKeyRepository;
 import com.example.spot.auth.domain.verification.VerificationCodeRepository;
 import com.example.spot.common.security.utils.JwtTokenProvider;
 import com.example.spot.common.security.utils.MemberUtils;
@@ -43,7 +43,6 @@ import java.security.SecureRandom;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
-import java.util.Objects;
 import java.util.Random;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
