+++ src/main/java/com/example/spot/auth/presentation/controller/legacy/AuthController.java
@@ -4,7 +4,7 @@ import com.example.spot.common.api.ApiResponse;
 import com.example.spot.common.api.code.status.SuccessStatus;
 import com.example.spot.common.security.utils.SecurityUtils;
 import com.example.spot.member.presentation.dto.MemberResponseDTO.NicknameDuplicateDTO;
-import com.example.spot.member.presentation.dto.rsa.Rsa;
+import com.example.spot.auth.presentation.dto.rsa.Rsa;
 import com.example.spot.auth.application.legacy.AuthService;
 import com.example.spot.common.presentation.validator.TextLength;
 import com.example.spot.member.presentation.dto.MemberRequestDTO;
